last one homie

+++ database/migrations/2025_06_27_100207_create_voedselpakket_table.php
@@ -8,15 +8,17 @@ class CreateVoedselpakketTable extends Migration
 {
     public function up()
     {
-        Schema::create('voedselpakket', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('klant_id')->nullable(); // nullable zodat 0 of null mag
-            $table->date('datum_samenstelling');
-            $table->date('datum_uitgifte')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('voedselpakket')) {
+            Schema::create('voedselpakket', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('klant_id')->nullable(); // nullable zodat 0 of null mag
+                $table->date('datum_samenstelling');
+                $table->date('datum_uitgifte')->nullable();
+                $table->timestamps();
 
-            $table->foreign('klant_id')->references('id')->on('klanten')->onDelete('cascade');
-        });
+                $table->foreign('klant_id')->references('id')->on('klanten')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
